Fix expanded not cancelling auto-close timers & add stagger behavior
